services/workers/requirements.txt
# Created automatically by Cursor AI (2024-08-27)

# Core dependencies
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0

# Redis and messaging
redis==5.0.1
nats-py==2.3.0

# ML and data processing
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
lightgbm==4.1.0
pyod==1.1.0
prophet==1.1.4
statsmodels==0.14.0

# Vector embeddings
pgvector==0.2.4
sentence-transformers==2.2.2

# PDF processing
pypdf2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10

# File processing
openpyxl==3.1.2
xlrd==2.0.1
python-dateutil==2.8.2

# HTTP and API
httpx==0.25.2
aiofiles==23.2.1

# AWS/S3
boto3==1.34.0

# Monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
